5GHz band ON/OFF from inferences results: * Mlp inference manager * Code refactor in wifi manager * Counters in separated class * Recipe update * Remove probability check in inferences * 5GHz band on/off * Counters management when band is off
