+++ files/server/managers/mlp_inference_manager/__init__.py
@@ -1,0 +1,2 @@
+"""MLP inference package"""
+from .service import mlp_inference_manager_service
