+++ files/server/managers/wifi_5GHz_band_manager/__init__.py
@@ -1,2 +1,3 @@
 """Wifi bands managment package"""
 from .service import wifi_5GHz_band_manager_service
+from .common import COUNTERS_ARRAY_SIZE_TO_PERFORM_INFERENCE
